feture/TEAMUP-40 добавил рубильники к эндпоинтам в ru.team.up.input.controller.publicController

See merge request survivalnightgame/teamup-core!132

+++ teamup-input/src/main/java/ru/team/up/input/controller/publicController/EventRestControllerPublic.java
@@ -58,6 +58,10 @@ public class EventRestControllerPublic {
     @GetMapping
     public EventDtoListResponse getAllEvents() {
         log.debug("Получение запроса на список мероприятий");
+        if (!ParameterService.getAllEventsEnabled.getValue()) {
+            log.debug("Метод getAllEvents выключен параметром getAllEventsEnabled = false");
+            throw new RuntimeException("Method getAllEvents is disabled by parameter getAllEventsEnabled");
+        }
         EventDtoListResponse eventDtoListResponse = EventDtoListResponse.builder().eventDtoList(
                         EventMapper.INSTANCE.mapDtoEventToEvent(eventServiceRest.getAllEvents()))
                 .build();
@@ -108,7 +112,10 @@ public class EventRestControllerPublic {
     @GetMapping(value = "/city/{city}")
     public EventDtoListResponse getAllEventByCity(@PathVariable String city) {
         log.debug("Запрос на поиск мероприятий по городу city: {}", city);
-
+        if (!ParameterService.getAllEventByCityEnabled.getValue()) {
+            log.debug("Метод getAllEventByCity выключен параметром getAllEventByCityEnabled = false");
+            throw new RuntimeException("Method getAllEventByCity is disabled by parameter getAllEventByCityEnabled");
+        }
         EventDtoListResponse eventDtoListResponse = EventDtoListResponse.builder().eventDtoList(
                         EventMapper.INSTANCE.mapDtoEventToEvent(eventServiceRest.getAllEventsByCity(city)))
                 .build();
@@ -132,6 +139,10 @@ public class EventRestControllerPublic {
     @GetMapping(value = "/name/{eventName}")
     public EventDtoListResponse findEventsByName(@PathVariable("eventName") String eventName) {
         log.debug("Получен запрос на поиск мероприятий по названию {}", eventName);
+        if (!ParameterService.getFindEventsByNameEnabled.getValue()) {
+            log.debug("Метод findEventsByName выключен параметром getfindEventsByNameEnabled = false");
+            throw new RuntimeException("Method findEventsByName is disabled by parameter getfindEventsByNameEnabled");
+        }
 
         EventDtoListResponse eventDtoListResponse = EventDtoListResponse.builder().eventDtoList(
                         EventMapper.INSTANCE.mapDtoEventToEvent(eventServiceRest.getEventByName(eventName)))
@@ -156,6 +167,10 @@ public class EventRestControllerPublic {
     @GetMapping(value = "/author")
     public ResponseEntity<List<Event>> findEventsByAuthor(@RequestBody UserRequest author) {
         log.debug("Получен запрос на поиск мероприятий по автору {}", author);
+        if (!ParameterService.getFindEventsByAuthorEnabled.getValue()) {
+            log.debug("Метод findEventsByAuthor выключен параметром getFindEventsByAuthorEnabled = false");
+            throw new RuntimeException("Method findEventsByAuthor is disabled by parameter getFindEventsByAuthorEnabled");
+        }
         List<Event> events = eventServiceRest.getAllEventsByAuthor(author.getUser().getId());
         Map<String, Object> monitoringParameters = new HashMap<>();
         if (events.isEmpty()) {
@@ -181,6 +196,10 @@ public class EventRestControllerPublic {
     @GetMapping(value = "/type")
     public ResponseEntity<List<Event>> findEventsByType(@RequestBody EventType eventType) {
         log.debug("Получен запрос на поиск мероприятий по типу: {}", eventType);
+        if (!ParameterService.getFindEventsByTypeEnabled.getValue()) {
+            log.debug("Метод findEventsByType выключен параметром getFindEventsByTypeEnabled = false");
+            throw new RuntimeException("Method findEventsByType is disabled by parameter getFindEventsByTypeEnabled");
+        }
         List<Event> events = eventServiceRest.getAllEventsByEventType(eventType);
         Map<String, Object> monitoringParameters = new HashMap<>();
         if (events.isEmpty()) {
@@ -207,6 +226,10 @@ public class EventRestControllerPublic {
     @PostMapping(value = "/")
     public ResponseEntity<Event> createEvent(@RequestBody EventRequest event) {
         log.debug("Получен запрос на создание мероприятия:\n {}", event);
+        if (!ParameterService.getCreateEventEnabled.getValue()) {
+            log.debug("Метод createEvent выключен параметром getCreateEventEnabled = false");
+            throw new RuntimeException("Method createEvent is disabled by parameter getCreateEventEnabled");
+        }
 
         checkEvent(event);
 
@@ -227,6 +250,10 @@ public class EventRestControllerPublic {
     @PutMapping(value = "/{id}")
     public ResponseEntity<Event> updateEvent(@RequestBody EventRequest event, @PathVariable("id") Long eventId) {
         log.debug("Получен запрос на обновление мероприятия {}", event);
+        if (!ParameterService.getUpdateEventEnabled.getValue()) {
+            log.debug("Метод updateEvent выключен параметром getUpdateEventEnabled = false");
+            throw new RuntimeException("Method updateEvent is disabled by parameter getUpdateEventEnabled");
+        }
 
         checkEvent(event);
 
@@ -246,6 +273,10 @@ public class EventRestControllerPublic {
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Event> deleteEvent(@PathVariable("id") Long eventId) {
         log.debug("Получен запрос на удаление мероприятия с id: {}", eventId);
+        if (!ParameterService.getDeleteEventEnabled.getValue()) {
+            log.debug("Метод deleteEvent выключен параметром getDeleteEventEnabled = false");
+            throw new RuntimeException("Method deleteEvent is disabled by parameter getDeleteEventEnabled");
+        }
         Event event = eventServiceRest.getEventById(eventId);
 
         if (event == null) {
@@ -269,6 +300,10 @@ public class EventRestControllerPublic {
     @PostMapping(value = "/join")
     public ResponseEntity<Event> addEventParticipant(@RequestBody JoinRequest joinRequest) {
         log.debug("Получен запрос на добавление участника мероприятия");
+        if (!ParameterService.getAddEventParticipantEnabled.getValue()) {
+            log.debug("Метод addEventParticipant выключен параметром getAddEventParticipantEnabled = false");
+            throw new RuntimeException("Method addEventParticipant is disabled by parameter getAddEventParticipantEnabled");
+        }
         Event event = eventServiceRest.addParticipant(joinRequest.getEventId(), joinRequest.getUserId());
 
         log.debug("Участник успешно добавлен");
@@ -285,6 +320,10 @@ public class EventRestControllerPublic {
     @PostMapping("/unjoin")
     public ResponseEntity<Event> deleteEventParticipant(@RequestBody JoinRequest joinRequest) {
         log.debug("Получен запрос на удаление участника мероприятия");
+        if (!ParameterService.getDeleteEventParticipantEnabled.getValue()) {
+            log.debug("Метод deleteEventParticipant выключен параметром getDeleteEventParticipantEnabled = false");
+            throw new RuntimeException("Method deleteEventParticipant is disabled by parameter getDeleteEventParticipantEnabled");
+        }
         Event event = eventServiceRest.deleteParticipant(joinRequest.getEventId(), joinRequest.getUserId());
 
         log.debug("Участник успешно удален");
